[CodeCompletion] Move CodeCompletionContext.h to its own file
