+++ include/swift/IDE/CodeCompletion.h
@@ -20,6 +20,7 @@
 #include "swift/Basic/OptionSet.h"
 #include "swift/Basic/StringExtras.h"
 #include "swift/Frontend/Frontend.h"
+#include "swift/IDE/CodeCompletionContext.h"
 #include "swift/IDE/CodeCompletionResult.h"
 #include "swift/IDE/CodeCompletionResultSink.h"
 #include "swift/IDE/CodeCompletionString.h"
@@ -93,87 +94,6 @@ public:
     if (I == depths.end())
       return None;
     return I->getValue();
-  }
-};
-
-class CodeCompletionContext {
-  friend class CodeCompletionResultBuilder;
-
-  /// A set of current completion results.
-  CodeCompletionResultSink CurrentResults;
-
-public:
-  CodeCompletionCache &Cache;
-  CompletionKind CodeCompletionKind = CompletionKind::None;
-
-  enum class TypeContextKind {
-    /// There is no known contextual type. All types are equally good.
-    None,
-
-    /// There is a contextual type from a single-expression closure/function
-    /// body. The context is a hint, and enables unresolved member completion,
-    /// but should not hide any results.
-    SingleExpressionBody,
-
-    /// There are known contextual types, or there aren't but a nonvoid type is expected.
-    Required,
-  };
-
-  TypeContextKind typeContextKind = TypeContextKind::None;
-
-  /// Whether there may be members that can use implicit member syntax,
-  /// e.g. `x = .foo`.
-  bool MayUseImplicitMemberExpr = false;
-
-  /// Flag to indicate that the completion is happening reusing ASTContext
-  /// from the previous completion.
-  /// NOTE: Do not use this to change the behavior. This is only for debugging.
-  bool ReusingASTContext = false;
-
-  CodeCompletionContext(CodeCompletionCache &Cache)
-      : Cache(Cache) {}
-
-  void setAnnotateResult(bool flag) { CurrentResults.annotateResult = flag; }
-  bool getAnnotateResult() const { return CurrentResults.annotateResult; }
-
-  void setIncludeObjectLiterals(bool flag) {
-    CurrentResults.includeObjectLiterals = flag;
-  }
-  bool includeObjectLiterals() const {
-    return CurrentResults.includeObjectLiterals;
-  }
-
-  void setAddInitsToTopLevel(bool flag) {
-    CurrentResults.addInitsToTopLevel = flag;
-  }
-  bool getAddInitsToTopLevel() const {
-    return CurrentResults.addInitsToTopLevel;
-  }
-
-  void setCallPatternHeuristics(bool flag) {
-    CurrentResults.enableCallPatternHeuristics = flag;
-  }
-  bool getCallPatternHeuristics() const {
-    return CurrentResults.enableCallPatternHeuristics;
-  }
-
-  void setAddCallWithNoDefaultArgs(bool flag) {
-    CurrentResults.addCallWithNoDefaultArgs = flag;
-  }
-  bool addCallWithNoDefaultArgs() const {
-    return CurrentResults.addCallWithNoDefaultArgs;
-  }
-
-  /// Allocate a string owned by the code completion context.
-  StringRef copyString(StringRef Str);
-
-  /// Sort code completion results in an implementation-defined order
-  /// in place.
-  static std::vector<CodeCompletionResult *>
-  sortCompletionResults(ArrayRef<CodeCompletionResult *> Results);
-
-  CodeCompletionResultSink &getResultSink() {
-    return CurrentResults;
   }
 };
