+++ lib/IDE/CMakeLists.txt
@@ -2,6 +2,7 @@
 add_swift_host_library(swiftIDE STATIC
   CodeCompletion.cpp
   CodeCompletionCache.cpp
+  CodeCompletionContext.cpp
   CodeCompletionDiagnostics.cpp
   CodeCompletionResultBuilder.cpp
   CodeCompletionResultPrinter.cpp
